Add classes for Agent, Credential, and Resource
These represent the fundamental concepts and tokens in the
Permit/Authority model. It's not entirely clear whether there should be
a base class here, since they all have the same interface, but I am
resisting that for now. There are reasonable doc strings on these,
though some of the content may belong in the overview.

This adds a YARD configuration, as well.