Add PermissionCheck and RoleCheck
These classes capture the authority and serve as self-evaluating rules.
The plan is that there can be a trivial factory available to policies
to pass in the user, resource, and authority, giving good ergonomics
without sacrificing design. For example, a can? method on a base policy
would create a PermissionCheck and evaluate it.