Set up Sequel connection and Permit model
- Add DB module
- Put initialization at the module/singleton level for now
- Add Permit model
- Build out Permits a little bit as a wrapper/repository

This really paves the way for the database interaction for the rest of
the modeling. Since the outside users will be using the Authority rather
than Permits directly, it may be an extraneous layer for now. I am kind
of playing with the ideas and just don't want Permit to bloat. It may be
that having a few different types to handle the searches is a better
design than a repository that basically just delegates. I'm picking up a
data clump smell with the (agents, credentials, resources, [zone])
parameter list, so that might be the first place to try extracting a
search concept.