Rework ResourceResolver to make Resources
Now that there are domain objects, the resolvers need to return them so
the other components, especially the Authority, can deal with its inputs
consistently.

This doesn't change the conceptual behavior of the ResourceResolver; it
still resolves an entity into a list of resources, including the entity
itself and a wildcard for all entities of its type.

Having a special resource that represents "all/any resources" allows the
conceptual model to hold together better than allowing nulls for either
the resource type or id. This was a big design question for some time,
but I consider it resolved now. All permits must have values for all
columns to keep the set operations working cleanly across zones.

- Force Resource type and id to strings
- Implement Resource equality
- Add Resource::ALL constant for the reserved "(all)" string
- Add Resource.all for getting the special "all resources" matcher