Build out actor-agent conversion
This is a pretty big change to how the AgentResolver works. The notion
of the UserDirectory has been removed, at least for now, and the
resolver implementation is now trivial. There are two important points
here:

1. The AgentResolver was originally planned to have more substantial
   implementation such that you would plug in a UserDirectory to
   extend it. Now, it simple enough that it could be subtituted
   directly, without another indirection point.
2. The details of account-type, affiliation, and so on are likely going
   to move to Keycard. For now, the resolver only returns the one agent,
   according to the default conversion.

The default conversion is now implemented in a separate class with the
semantics and details documented and tested. Agent could be subclassed
with ::conversion defined to supply a different implementation, though
it has extension built into its design. It allows the actor object to
implement #to_agent, as well as a fall-through to ask the agent_type and
agent_id, then to the class name and id properties.