Reorganize resolvers and conversions
This is a larger diff I'd like, since moving things around obscures some
of the functional/design changes. They are as follows:

- Change the CredentialResolver to return real Credentials, not strings
- Extract resource conversion fromm the ResourceResolver
- Split resource conversion into entity (base) and wildcard
- Eliminate `#type` as a method that can be sent during resource conversion
  since it is overloaded
- Make sure that Credential coerces to strings
- Tighten up a few specs as far as error expectations
- Check for the COVERAGE or TRAVIS envvars before running Coveralls

There is also a pretty substantial reorgnanization of the source files
and requires in checkpoint.rb, centering around agent, credential, and
resource each pulling in their related files, and making directories for
items corresponding to each key concept.