Convert checks to predicate queries
- Add Checkpoint::Query namespace
- Change RoleCheck to RoleGranted and move to Query
- Change PermissionCheck ActionPermitted and move to Query

The top-level design is now straining a little in that Agent,
Credential, and Resource are currently really the token types. Where
this is showing up the most is in the resolvers. Specifically, I had
to pass in Resource and Credential objects from the outside to drive a
meaningful integration test of granting and testing for a role at the
system level. These are now type checked in the resolution, which is
rather smelly.

Now that there are namespaces for each of top-level concepts, I think
the token behavior should be pushed down and allow the base names to
serve as wrappers/decorators for the actual objects. What I think that
will afford us is the ability to make Permission and Role types that are
not tokens, but can be converted to tokens as polymorphic Credentials,
and simplify the credential resolution to some degree. This is all
hinging on the need to treat roles and permissions differently in some
places, but wanting to treat them generically for the permit/grant
model. The PermissionMapper will also be called into question during
that change -- the resolver will be simplified to the degree that it
will likely be better to just allow an alternate implementation (as
happened with the AgentResolver).