Fix the tests that were expecting a string version of token
There was one spot in the permits spec that was creating a Credential
explicitly where it needed a token. Fixing that removed the need for the
`_token` method since now `#token` was working in the only spot it
was breaking.

Also add defaults to the constructor parameters optional to get our two
tests closer to the shape we want (only name), so we can go back and
forth to red by changing the constructor and find callsites that need to
use the new Credential instead. This commit is in broken mode to show
the list (16 failures).

I also recorded a quick Vim macro to uncomment the current line, move to
the other, comment it, and save to auto rerun the tests. The same could
be done with undo/redo/save on a single line. I use vim-commentary, so
recording was "qt gcc ddp k gcc :w<ENTER> q". Then @t to run it, and @@ to
rerun.