Remove the named parameter -- the refactoring is done
The refactoring is complete, but there are needed tests and comments.
For example, the eql? and == tests don't check for type, which they
probably should. The new classes also need doc comments and the rest
need a once-over to deal with the change in the resolver, primarily. It
is now dealing with richer objects, so the docs need to be revised.

The PermissionMapper has now also outlived its usefulness. Customizing
the resolver now only requires reimplementing the one method.