Finally sort out migrations and startup with Rails
It was quite difficult to sort out all of the tricky edge cases and
timing issues with the Rails initialization process vs. Rake, but I am
very happy with the result now. The DB module is more convenient and
better behaved, and there are lots of inline docs about the timing
specifics in the Railtie and Rake task.

This commit also adds some critical support for Rails apps in that
db:setup will work properly, setting the checkpoint schema version, and
db:migrate will dump it. Rails includes every table in the schema
(matching the prefix, I'm sure), so the Checkpoint tables end up in
schema.rb. This is fine and convenient, but Rails doesn't know that the
schema_info value has to be updated, and Sequel migrations would be
attempted again. This was the source of the most recent CI build
failures.