CP-38 - Support agent conversion
In addition to expansion, this adds support at the resolver for single
conversion. It removes the `.from` factory method from agent and moves
it to the resolver, which is a rather natural place for it. This also
renames the resolve method, which is a breaking API change for any
subclasses (as in KCV).