CP-40 - Support resource conversion
Like implemented for agents, this change supports conversion of
application entities to Resources. There was some additional
simplification at the Resource and wildcard types, where the factory
methods don't really make sense with a conversion defined on the
resolver and having a convenient way to promote a Resource to a wildcard
of that type. Removing the .from methods eliminates some duplicative
logic and tests around conversion.