CP-39 - Rework the Credential Resolver
This is a pretty significant change to how the Credential Resolver
works. By splitting the concept of conversion and expansion, the design
was driven to have a permission-only base resolver with an alternative
role-map resolver for applications needing basic mapping.

This change eliminates the PermissionMapper, normalizing the design with
the other resolvers somewhat.