CP-42 - Add to_agent and to_resource methods
To normalize the Agent and Resources interfaces with that of Credential,
this adds Agent#to_agent and Resource#to_resource.

This is driven by the change where conversion was moved from these
classes to the resolvers. As an example of where this comes up, passing
Resource.all to the new Authority#permit! method was passing it along to
convert, which considered the Resource::AllOfAnyType object as an
application object to be wrapped in a new Resource (getting the awkward
type of Checkpoint::Resource::AllOfAnyType). Rather than do specific
type-checking, we check if it responds to #to_resource, which, as of
this commit, it will, returning itself, giving the appropriate type/id.