CP-42 - Add who, what, which to Permits
The who, what, and which methods on Permits provide a low-level
interface on the repository for searching for grants along two of the
dimensions and returning a set on the third. These are well suited to
finding agents, credentials, and resources respectively. They each
return the full grants, to be converted and deduplicated by a higher
level object like the authority.

As an example of this deduplication, a user may have both a permission
and a role that applies for a given resource, so there would be two
grants. The repository should return both of them, but the authority
should return only the one agent.