CP-42 - Use type and id for == on Agent and Resource
Agent and Resource were comparing for equality by delegating to the
actor and entity, respectively. This is a small change to disregard the
actual object type of the concrete object and compare the agent or
resource by value; specifically, only the type and id are compared.

In practice, it will be rare that two concrete objects that report the
same type and id, but not report == equality with each other, but this
supports testing (doubles without stubbed responses to ==) and scenarios
where it is not practical to modify a class to recognize another as ==,
but they should be considered equal for agent or resource identification
purposes.

The stricter equality of eql? still delegates to the underlying objects.