CP-42 - Add who to Authority
This finishes off the ability to query for agents who can take a given
action on a given resource. The who method returns tokens rather than
Agents because we do not have a concrete actor to wrap.

We will have to do some experimentation to see whether the resolvers
should have a token-to-domain method, or if that should reside more
fully on the app side.