CP-41 - Rename "permit" to "grant"
This change reverts the naming experiment of trying to use "permit" as
the noun form. We have found that it did not really solve the issue of
distinguishing between the noun and verb, and "grant" is a more common
term. The one exception to where grant did not replace permit is in the
Authority, where the question is whether the authority permits an
action. It does not seem natural to ask whether an authority grants
a credential, nor does it seem particularly natural to ask the authority
whether a credential is granted -- so neither grants? nor granted? are
used here.