CP-53 - Add convenient methods for testing Resource wildcards
The #all? and #all_of_type? methods on Resource and Resource::Token make
it easy to check whether or not the object held is a wildcard for all
entities of a type or all of any type. The predicate methods use the
same base names as those used to create wildcards from entities.