CP-54 - Add aliases so tokens can convert and be granted
At some point, we stopped calling #type for conversion, which I think is
still probably correct. This change adds type-specific aliases like
agent_type to type on Agent::Token so the tokens can be granted without
having a reference to the real entity. This is the reason they were
separated off in the first place, so this was just a plain-old bug.